requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
audioop-lts==0.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
dotenv==0.9.9
fastapi==0.122.0
gTTS==2.5.4
h11==0.16.0
idna==3.11
numpy==2.3.5
paho-mqtt==2.1.0
pandas==2.3.3
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
